requirements.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
bidict==0.21.4
blinker==1.4
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
charset-normalizer==2.0.10
clang==5.0
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==2.0.1
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
flatbuffers==1.12
foolnltk==0.1.7
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.3
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-engineio==4.3.1
python-socketio==5.5.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.15.0
snownlp==0.12.3
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.2.0
zipp==3.6.0
